Created basic code for database interaction
It took way too long, but I finally have the basic code to communicate with the
database in a useful way without a lot of code repetition. Instead of
SQLalchemy, we deciced to use SQLModel since that plays nicely with
Pydantic models. We creates a context manager for `Context`-objects,
which represent a specific context in which something gets done within
the database. We also created unit tests.

A lot of work still has to be done, but the tracing bullets are starting
to hit the target!