Changed the requirement so unit tests can pass
Before this commit, the `requirements.txt` needed a production ready
version of the `my-model` package. In the last production ready version
is a bug however; the `password_hash` for users is required. This is
fixed in the latest `dev` release.

Our unit tests fail because of this. So we are going to use the `dev`
release for the time being. When going to production, we have to make
sure only production versions are used. So we cannot go to production
until this bug is fixed.